fix: conflict when rebase from origin/main

+++ src/main/java/vn/id/nonglam/kltn/kltn/dto/response/hotel/HotelResponse.java
@@ -2,6 +2,7 @@ package vn.id.nonglam.kltn.kltn.dto.response.hotel;
 
 import vn.id.nonglam.kltn.kltn.common.enums.HotelStatus;
 
+import java.math.BigDecimal;
 import java.util.Set;
 import java.util.UUID;
 
@@ -14,8 +15,8 @@ public record HotelResponse(UUID id, String name, int countComments, double avgR
     public record RoomUtilityResponse(String name, String iconCode) {}
 
     public record RoomTypeResponse(UUID id, String name, String description, int capacity,
-                                    double price, Set<RoomUtilityResponse> roomUtilities, 
-                                    Set<String> roomTypeImages, double depositedPercent) {}
+                                   BigDecimal price, Set<RoomUtilityResponse> roomUtilities,
+                                   Set<String> roomTypeImages, double depositedPercent) {}
 
     public record AddressResponse(UUID id, String street, String ward, String province, int postalCode, double latitude, double longitude) {}
